fix(combobox): add checking open state for dropdown content
