+++ packages/react-magma-dom/src/components/Combobox/Combobox.tsx
@@ -1,5 +1,11 @@
 import * as React from 'react';
 
+import {
+  AlignedPlacement,
+  autoUpdate,
+  flip,
+  useFloating,
+} from '@floating-ui/react-dom';
 import { useCombobox } from 'downshift';
 import { CloseIcon } from 'react-magma-icons';
 
@@ -24,7 +30,6 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
     defaultSelectedItem,
     disableCreateItem,
     errorMessage,
-    floatingElementStyles,
     hasError,
     helperMessage,
     initialSelectedItem,
@@ -57,8 +62,6 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
     onItemCreated,
     placeholder,
     selectedItem: passedInSelectedItem,
-    setFloating,
-    setReference,
     toggleButtonRef,
   } = props;
 
@@ -271,6 +274,23 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
       onInputKeyDown(event);
   }
 
+  const { floatingStyles, refs, elements, update } = useFloating({
+    middleware: [flip()],
+    placement: 'bottom-start' as AlignedPlacement,
+    whileElementsMounted: autoUpdate,
+  });
+
+  React.useEffect(() => {
+    const referenceComboboxInput = elements.reference;
+    const floatingItemsList = elements.floating;
+
+    if (isOpen && referenceComboboxInput && floatingItemsList) {
+      return autoUpdate(referenceComboboxInput, floatingItemsList, update);
+    }
+  }, [isOpen, elements, update]);
+
+  const floatingElementStyles = { ...floatingStyles, width: '100%' };
+
   return (
     <SelectContainer
       descriptionId={ariaDescribedBy}
@@ -313,7 +333,7 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
         onInputKeyPress={onInputKeyPress}
         onInputKeyUp={onInputKeyUp}
         placeholder={placeholder}
-        setReference={setReference}
+        setReference={refs.setReference}
         toggleButtonRef={toggleButtonRef}
       >
         {isClearable && selectedItem && (
@@ -341,7 +361,7 @@ export function InternalCombobox<T>(props: ComboboxProps<T>) {
         isLoading={isLoading && isTypeahead}
         maxHeight={itemListMaxHeight || theme.combobox.menu.maxHeight}
         menuStyle={menuStyle}
-        setFloating={setFloating}
+        setFloating={refs.setFloating}
       />
     </SelectContainer>
   );
