+++ packages/react-magma-dom/src/components/Combobox/MultiCombobox.tsx
@@ -1,5 +1,11 @@
 import * as React from 'react';
 
+import {
+  AlignedPlacement,
+  autoUpdate,
+  flip,
+  useFloating,
+} from '@floating-ui/react-dom';
 import { useCombobox, useMultipleSelection } from 'downshift';
 import { CloseIcon } from 'react-magma-icons';
 
@@ -25,7 +31,6 @@ export function MultiCombobox<T>(props: MultiComboboxProps<T>) {
     defaultItems,
     disableCreateItem,
     errorMessage,
-    floatingElementStyles,
     hasError,
     hasPersistentMenu = false,
     helperMessage,
@@ -57,8 +62,6 @@ export function MultiCombobox<T>(props: MultiComboboxProps<T>) {
     onIsOpenChange,
     onItemCreated,
     onRemoveSelectedItem,
-    setReference,
-    setFloating,
     placeholder,
     toggleButtonRef,
   } = props;
@@ -368,6 +371,23 @@ export function MultiCombobox<T>(props: MultiComboboxProps<T>) {
       onInputKeyDown(event);
   }
 
+  const { floatingStyles, refs, elements, update } = useFloating({
+    middleware: [flip()],
+    placement: 'bottom-start' as AlignedPlacement,
+    whileElementsMounted: autoUpdate,
+  });
+
+  React.useEffect(() => {
+    const referenceMultiComboboxInput = elements.reference;
+    const floatingItemsList = elements.floating;
+
+    if (isOpen && referenceMultiComboboxInput && floatingItemsList) {
+      return autoUpdate(referenceMultiComboboxInput, floatingItemsList, update);
+    }
+  }, [isOpen, elements, update]);
+
+  const floatingElementStyles = { ...floatingStyles, width: '100%' };
+
   return (
     <SelectContainer
       descriptionId={ariaDescribedBy}
@@ -411,7 +431,7 @@ export function MultiCombobox<T>(props: MultiComboboxProps<T>) {
         onInputKeyUp={onInputKeyUp}
         placeholder={selectedItems.length > 0 ? null : placeholder}
         selectedItems={selectedItemsContent}
-        setReference={setReference}
+        setReference={refs.setReference}
         toggleButtonRef={toggleButtonRef}
       >
         {isClearable && selectedItems?.length > 0 && (
@@ -440,7 +460,7 @@ export function MultiCombobox<T>(props: MultiComboboxProps<T>) {
         isLoading={isLoading && isTypeahead}
         maxHeight={itemListMaxHeight || theme.combobox.menu.maxHeight}
         menuStyle={menuStyle}
-        setFloating={setFloating}
+        setFloating={refs.setFloating}
       />
     </SelectContainer>
   );
